+++ module-api/src/main/java/com/gdsc/domain/user/model/UserRequest.java
@@ -2,10 +2,14 @@ package com.gdsc.domain.user.model;
 
 
 import com.gdsc.domain.user.entity.Gender;
+import jakarta.validation.constraints.NotBlank;
 
-public record UserInfoRequest(
+public record UserRequest(
+        @NotBlank
         String nickname,
+        @NotBlank
         Gender gender,
+        @NotBlank
         Integer age
 ) {
 }
