+++ module-api/src/main/java/com/gdsc/domain/user/service/UserService.java
@@ -2,9 +2,8 @@ package com.gdsc.domain.user.service;
 
 import com.gdsc.domain.user.entity.User;
 import com.gdsc.security.service.UserDetailService;
-import com.gdsc.domain.user.model.UserInfoRequest;
+import com.gdsc.domain.user.model.UserRequest;
 
-import com.gdsc.domain.user.model.UserKeywordsRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,14 +14,9 @@ public class UserService {
 
     private final UserDetailService userDetailService;
     @Transactional
-    public User updateUserInfo(UserInfoRequest userInfoRequest, User user) {
+    public User updateUserInfo(UserRequest userRequest, User user) {
         User findUser = userDetailService.loadUserByUsername(user.getFirebaseUid());
-        return findUser.updateUser(userInfoRequest.nickname(), userInfoRequest.gender(), userInfoRequest.age());
-    }
 
-    @Transactional
-    public User updateKeywords(UserKeywordsRequest userKeywordsRequest, User user){
-        User findUser = userDetailService.loadUserByUsername(user.getFirebaseUid());
-        return findUser.updateKeywords(userKeywordsRequest.keyword1(), userKeywordsRequest.keyword2(), userKeywordsRequest.keyword3());
+        return findUser.updateUser(userRequest.nickname(), userRequest.gender(), userRequest.age());
     }
 }
